# api/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Graphic Designer API
  description: |
    # Smart Graphic Designer API
    
    A comprehensive AI-powered graphic design generation service that creates professional designs
    based on text prompts, brand constraints, and reference materials.
    
    ## Key Features
    
    - **AI-Powered Generation**: Uses state-of-the-art models like Gemini 2.5 Flash Image
    - **Brand Compliance**: Enforces color palettes, typography, and logo safe zones
    - **Multiple Formats**: Supports PNG, JPG, and WebP outputs
    - **Quality Assurance**: Built-in guardrails and validation
    - **Real-time Updates**: WebSocket support for progress tracking
    - **Comprehensive Audit**: Full traceability and cost tracking
    
    ## Authentication
    
    This API uses JWT authentication via Kong API Gateway.
    
    ## Rate Limits
    
    - **Standard**: 100 requests per minute per API key
    - **Burst**: 30 requests in quick succession
    
    ## Cost Structure
    
    | Operation | Estimated Cost |
    |-----------|----------------|
    | Simple Design | $0.02-0.05 |
    | Complex Design | $0.05-0.10 |
    | Bulk Generation | $0.15-0.30 |
    
  version: 1.0.0
  contact:
    name: SGD API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com
    description: Production server
  - url: http://localhost:8000
    description: Local development server
security:
  - JWTAuth: []
paths:
...
  /canon/derive:
    post:
      summary: Derive Brand Canon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanonDeriveRequest'
      responses:
        '200':
          description: Canon
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CanonDeriveResponse'
...
components:
  schemas:
...
    CanonDeriveRequest:
      type: object
      required: [project_id, evidence_ids]
      properties:
        project_id: { type: string }
        evidence_ids: { type: array, items: { type: string } }
    CanonDeriveResponse:
      type: object
      properties:
        palette_hex: { type: array, items: { type: string } }
        fonts: { type: array, items: { type: string } }
        voice:
          type: object
          properties:
            tone: { type: string }
            dos: { type: array, items: { type: string } }
            donts: { type: array, items: { type: string } }
...